Refactoring, type hinting and general code clean up to appease linters and human developers alike
