+++ caracara/client.py
@@ -141,9 +141,11 @@ class Client:
             self.api_authentication = falconpy_authobject
         else:
             raise Exception("Impossible authentication scenario")
+
         self.logger.info("Requesting API token")
         self.api_authentication.token()  # Need to force the authentication to resolve the base_url
         self.logger.info("Resolved Base URL: %s", self.api_authentication.base_url)
+
         # Configure modules here so that IDEs can pick them up
         self.logger.debug("Setting up Hosts module")
         self.hosts = HostsApiModule(self.api_authentication)
