+++ examples/common/exceptions.py
@@ -8,7 +8,11 @@ class BaseError(Exception):
 class MissingArgument(BaseError):
     """A required argument was not provided. Check YAML configuration."""
 
-    def __init__(self, argument_name, message="A required argument was not provided."):
+    def __init__(
+        self,
+        argument_name: str,
+        message: str = "A required argument was not provided.",
+    ):
         """Intialize a MissingArgument error."""
         self.argument_name = argument_name
         self.message = message
@@ -22,10 +26,11 @@ class MissingArgument(BaseError):
 class NoDevicesFound(BaseError):
     """No devices were found matching the filter provided."""
 
-    def __init__(self,
-                 filter_fql="No filters applied",
-                 message="No devices matched the provided Falcon Filter."
-                 ):
+    def __init__(
+        self,
+        filter_fql: str = "No filters applied",
+        message: str = "No devices matched the provided Falcon Filter.",
+    ):
         """Initialize a NoDevicesFound error."""
         self.filter_fql = filter_fql
         self.message = message
@@ -39,10 +44,11 @@ class NoDevicesFound(BaseError):
 class NoGroupsFound(BaseError):
     """No groups were found matching the filter provided."""
 
-    def __init__(self,
-                 filter_fql="No filters applied",
-                 message="No groups matched the provided Falcon Filter."
-                 ):
+    def __init__(
+        self,
+        filter_fql: str = "No filters applied",
+        message: str = "No groups matched the provided Falcon Filter.",
+    ):
         """Initialize a NoGroupsFound error."""
         self.filter_fql = filter_fql
         self.message = message
@@ -56,10 +62,11 @@ class NoGroupsFound(BaseError):
 class NoLoginsFound(BaseError):
     """No logins were found matching the filter provided."""
 
-    def __init__(self,
-                 filter_fql="Recent logins",
-                 message="No logins matched the provided Falcon Filter."
-                 ):
+    def __init__(
+        self,
+        filter_fql: str = "Recent logins",
+        message: str = "No logins matched the provided Falcon Filter.",
+    ):
         """Initialize a NoLoginsFound error."""
         self.filter_fql = filter_fql
         self.message = message
@@ -73,10 +80,11 @@ class NoLoginsFound(BaseError):
 class NoAddressesFound(BaseError):
     """No network address changes were found matching the filter provided."""
 
-    def __init__(self,
-                 filter_fql="Recent network addresses",
-                 message="No logins matched the provided Falcon Filter."
-                 ):
+    def __init__(
+        self,
+        filter_fql: str = "Recent network addresses",
+        message: str = "No logins matched the provided Falcon Filter.",
+    ):
         """Initialize a NoAddressesFound error."""
         self.filter_fql = filter_fql
         self.message = message
@@ -90,7 +98,10 @@ class NoAddressesFound(BaseError):
 class NoSessionsConnected(BaseError):
     """No successful connections were made for the RTR batch request."""
 
-    def __init__(self, message="No successful connections were made for this batch."):
+    def __init__(
+        self,
+        message: str = "No successful connections were made for this batch.",
+    ):
         """Initialize a NoSessionsConnected error."""
         self.message = message
         super().__init__(self.message)
