+++ pyproject.toml
@@ -44,6 +44,13 @@ disable = [
     # based on data returned by Falcon. It would be silly for us to refactor classes
     # representing these constructs just to pass this pylint rule. 
     "too-many-instance-attributes",
+
+    # Disable R0904 (too-many-public-methods) as we have some modules, like hosts, which
+    # by design have many functions exposed. As this is an SDK wrapper, we fully expect to
+    # implement most of the functionality of the underlying SDK (FalconPy) and APIs
+    # (from Falcon) by definition. Therefore, we have switched this rule off globally.
+    "too-many-public-methods",
+
     # Disable pylint rule R0201 as we have functions that are decorated with decorators
     # that expect a self parameter. In these cases, the functions cannot be decorated
     # with @staticmethod.
